GH-479 package renaming
