+++ github-integration/src/main/java/de/zalando/zally/integration/ValidationController.kt
@@ -1,0 +1,48 @@
+package de.zalando.zally.integration
+
+import de.zalando.zally.integration.config.logger
+import de.zalando.zally.integration.validation.ValidationService
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestHeader
+import org.springframework.web.bind.annotation.ResponseBody
+import org.springframework.web.bind.annotation.ResponseStatus
+import org.springframework.web.bind.annotation.RestController
+import javax.servlet.http.HttpServletRequest
+
+private val PULL_REQUEST_EVENT_NAME = "pull_request"
+
+@RestController
+class ApiValidationController(private val validationService: ValidationService) {
+
+    val log by logger()
+
+    @ResponseBody
+    @PostMapping("/github-webhook")
+    @ResponseStatus(value = HttpStatus.ACCEPTED)
+    fun validatePullRequest(@RequestBody payload: String,
+                            @RequestHeader(value = "X-GitHub-Event") eventType: String,
+                            @RequestHeader(value = "X-Hub-Signature") signature: String) {
+
+        if (eventType != PULL_REQUEST_EVENT_NAME) {
+            log.info("Received unsupported event: {}", eventType)
+            return
+        }
+
+        log.info("Received webhook: {}", eventType)
+
+        validationService.validatePullRequest(payload, signature)
+
+        log.info("Finished webhook processing")
+    }
+
+    @ExceptionHandler(Exception::class)
+    @ResponseBody
+    fun handleControllerException(request: HttpServletRequest, ex: Throwable): ResponseEntity<*> {
+        log.error("failed", ex)
+        return ResponseEntity<Any>(HttpStatus.INTERNAL_SERVER_ERROR)
+    }
+}
