+++ github-integration/src/main/java/de/zalando/zally/integration/zally/Violation.kt
@@ -1,4 +1,4 @@
-package de.zalando.zally.github.dto
+package de.zalando.zally.integration.zally
 
 data class Violation(
