+++ github-integration/src/test/java/de/zalando/zally/integration/TestUtil.kt
@@ -1,4 +1,4 @@
-package de.zalando.zally.github
+package de.zalando.zally.integration
 
 import org.apache.commons.io.IOUtils
